chore: now you can select repositories for a given organization
