+++ components/Workspaces/TrackedRepoWizard/TrackedRepoWizard.tsx
@@ -5,17 +5,24 @@ import { TrackedRepoWizardLayout } from "./TrackedRepoWizardLayout";
 import { SearchByReposStep } from "./SearchByReposStep";
 import { PasteReposStep } from "./PasteReposStep";
 import { FilterPastedReposStep } from "./FilterPastedReposStep";
-import { SearchReposByOrgStep } from "./SearchReposByOrgStep";
+import { SelectOrgReposStep } from "./SelectOrgReposStep";
 
 interface TrackedReposWizardProps {
   onAddToTrackingList: (repos: Map<string, boolean>) => void;
   onCancel: () => void;
 }
 
-type TrackedReposStep = "pickReposOrOrg" | "pickRepos" | "pasteRepos" | "pickOrg" | "filterPastedRepos";
+type TrackedReposStep =
+  | "pickReposOrOrg"
+  | "pickRepos"
+  | "pasteRepos"
+  | "pickOrg"
+  | "filterPastedRepos"
+  | "pickOrgRepos";
 
 export const TrackedReposWizard = ({ onAddToTrackingList, onCancel }: TrackedReposWizardProps) => {
   const [step, setStep] = useState<TrackedReposStep>("pickReposOrOrg");
+  const [organization, setOrganization] = useState<string>("open-sauced");
   const [currentTrackedRepositories, setCurrentTrackedRepositories] = useState<Map<string, boolean>>(new Map());
   const suggestedRepos: any[] = [];
   const [searchTerm, setSearchTerm] = useState<string | undefined>();
@@ -66,6 +73,7 @@ export const TrackedReposWizard = ({ onAddToTrackingList, onCancel }: TrackedRep
         break;
       default:
         setStep("pickReposOrOrg");
+        break;
     }
   }
 
@@ -118,17 +126,16 @@ export const TrackedReposWizard = ({ onAddToTrackingList, onCancel }: TrackedRep
             repositories={repositories}
           />
         );
-      // TODO: other steps
       case "pickOrg":
+        setStep("pickOrgRepos");
+        break;
+      case "pickOrgRepos":
         return (
-          <SearchReposByOrgStep
-            onSelectRepo={onSelectRepo}
+          <SelectOrgReposStep
+            repositories={repositories}
+            organization={organization}
             onToggleRepo={onToggleRepo}
             onToggleAllRepos={onToggleAllRepos}
-            onSearch={onSearchRepos}
-            repositories={repositories}
-            searchedRepos={searchedRepos}
-            suggestedRepos={suggestedRepos}
           />
         );
